+++ lib/config/apiConfig.ts
@@ -1,5 +1,3 @@
 import { ApiStrategy } from '@lib/constants/apiStrategy';
 
-const raw = process.env.NEXT_PUBLIC_API_STRATEGY;
-
-export const API_STRATEGY: ApiStrategy = raw === 'grpc' ? ApiStrategy.GRPC : ApiStrategy.REST;
+export const API_STRATEGY: ApiStrategy = process.env.NEXT_PUBLIC_API_STRATEGY === 'grpc' ? ApiStrategy.GRPC : ApiStrategy.REST;
