+++ lib/hooks/mutations/useDeleteTrackFileMutation.ts
@@ -1,22 +1,22 @@
-import { deleteTrackGrpc } from '@api/resources/trackGrpc';
-import { deleteTrack } from '@api/resources/trackRest';
+import { deleteTrackFileGrpc } from '@api/resources/trackGrpc';
+import { deleteTrackFile } from '@api/resources/trackRest';
 import { API_STRATEGY } from '@lib/config/apiConfig';
 import { ApiStrategy } from '@lib/constants/apiStrategy';
 import { tryCatch } from '@lib/utils/neverthrowUtils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-type DeleteTrackParams = {
-	id?: string;
+type DeleteTrackFileParams = {
+	id: string;
 	onSuccess?: () => void;
 	onError?: (error: Error) => void;
 };
 
-export function useDeleteTrack({ onSuccess, onError }: DeleteTrackParams) {
+export function useDeleteTrackFileMutation({ id, onSuccess, onError }: DeleteTrackFileParams) {
 	const queryClient = useQueryClient();
 
 	return useMutation({
-		mutationFn: async (id: string) => {
-			const result = API_STRATEGY === ApiStrategy.REST ? await deleteTrack(id) : await deleteTrackGrpc(id);
+		mutationFn: async () => {
+			const result = API_STRATEGY === ApiStrategy.REST ? await deleteTrackFile(id) : await deleteTrackFileGrpc(id);
 
 			if (result.isErr()) throw result.error;
 			return result.value;
@@ -24,12 +24,12 @@ export function useDeleteTrack({ onSuccess, onError }: DeleteTrackParams) {
 		onSuccess: async () => {
 			const invalidateResult = await tryCatch(() => queryClient.invalidateQueries({ queryKey: ['tracks'] }));
 			if (invalidateResult.isErr()) {
-				console.warn('Track deleted, but cache invalidation failed: ', invalidateResult.error);
+				console.warn('Track file deleted, but cache invalidation failed: ', invalidateResult.error);
 			}
 			onSuccess?.();
 		},
 		onError: (error) => {
-			console.error('Delete track error:', error);
+			console.error('Delete track file error:', error);
 			onError?.(error);
 		},
 	});
