+++ lib/hooks/mutations/useDeleteTrackMutation.ts
@@ -1,22 +1,22 @@
-import { deleteTrackFileGrpc } from '@api/resources/trackGrpc';
-import { deleteTrackFile } from '@api/resources/trackRest';
+import { deleteTrackGrpc } from '@api/resources/trackGrpc';
+import { deleteTrack } from '@api/resources/trackRest';
 import { API_STRATEGY } from '@lib/config/apiConfig';
 import { ApiStrategy } from '@lib/constants/apiStrategy';
 import { tryCatch } from '@lib/utils/neverthrowUtils';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-type DeleteTrackFileParams = {
-	id: string;
+type DeleteTrackParams = {
+	id?: string;
 	onSuccess?: () => void;
 	onError?: (error: Error) => void;
 };
 
-export function useDeleteTrackFile({ id, onSuccess, onError }: DeleteTrackFileParams) {
+export function useDeleteTrackMutation({ onSuccess, onError }: DeleteTrackParams) {
 	const queryClient = useQueryClient();
 
 	return useMutation({
-		mutationFn: async () => {
-			const result = API_STRATEGY === ApiStrategy.REST ? await deleteTrackFile(id) : await deleteTrackFileGrpc(id);
+		mutationFn: async (id: string) => {
+			const result = API_STRATEGY === ApiStrategy.REST ? await deleteTrack(id) : await deleteTrackGrpc(id);
 
 			if (result.isErr()) throw result.error;
 			return result.value;
@@ -24,12 +24,12 @@ export function useDeleteTrackFile({ id, onSuccess, onError }: DeleteTrackFilePa
 		onSuccess: async () => {
 			const invalidateResult = await tryCatch(() => queryClient.invalidateQueries({ queryKey: ['tracks'] }));
 			if (invalidateResult.isErr()) {
-				console.warn('Track file deleted, but cache invalidation failed: ', invalidateResult.error);
+				console.warn('Track deleted, but cache invalidation failed: ', invalidateResult.error);
 			}
 			onSuccess?.();
 		},
 		onError: (error) => {
-			console.error('Delete track file error:', error);
+			console.error('Delete track error:', error);
 			onError?.(error);
 		},
 	});
