+++ lib/hooks/queries/useTracksQuery.ts
@@ -5,14 +5,12 @@ import { useQuery } from '@tanstack/react-query';
 import { BaseResourceError } from '@models/errors/baseResourceError';
 import { API_STRATEGY } from '@lib/config/apiConfig';
 import { ApiStrategy } from '@lib/constants/apiStrategy';
-import { trackClient } from '@api/infrastructure/grpc/clients';
 import { fetchTracksGrpc } from '@api/resources/trackGrpc';
 
-export const useTracks = (filters: TracksRequestQueryParams) =>
+export const useTracksQuery = (filters: TracksRequestQueryParams) =>
 	useQuery({
 		queryKey: ['tracks', filters],
 		queryFn: async () => {
-			await trackClient.listTracks(filters);
 			const result = API_STRATEGY === ApiStrategy.REST ? await fetchTracks(filters) : await fetchTracksGrpc(filters);
 
 			if (result.isErr()) {
