+++ tests/integration/useTracksListController.logic.test.tsx
@@ -10,7 +10,7 @@ vi.mock('@components/tracksManager/tacksList/Columns', () => ({
 	getTrackColumns: vi.fn(),
 }));
 
-import { useTracks } from '@lib/hooks/fetchers/useFetchTracks';
+import { useTracksQuery } from '@lib/hooks/queries/useTracksQuery';
 import { useQueryTableFilters } from '@lib/hooks/components/trackList/useQueryTableFilters';
 import { getTrackColumns } from '@components/tracksManager/tacksList/Columns';
 import type { PaginationState } from '@tanstack/react-table';
@@ -29,7 +29,7 @@ describe('useTracksListController', () => {
 	beforeEach(() => {
 		vi.resetAllMocks();
 
-		(useTracks as Mock).mockReturnValue({
+		(useTracksQuery as Mock).mockReturnValue({
 			data: fakeTrackData,
 			isLoading: false,
 			error: null,
@@ -95,7 +95,7 @@ describe('useTracksListController', () => {
 			setPageQuery: mockSetPageQuery,
 		});
 
-		(useTracks as Mock).mockReturnValueOnce({
+		(useTracksQuery as Mock).mockReturnValueOnce({
 			data: { data: [], meta: { totalPages: 5 } },
 			isLoading: false,
 			error: null,
